Bypass chat conversion when uploading cha

+++ backend/analysis/convert/convert.py
@@ -9,20 +9,25 @@ def convert(transcript):
     transcript.save()
 
     try:
-        reader = SifReader(transcript.content.path)
-        cha_name = transcript.content.name.replace('.txt', '.cha')
-        cha_path = transcript.content.path.replace('.txt', '.cha')
-        # use title from metadata, if present
-        title = reader.document.title
-        if title:
-            cha_name = cha_name.replace(transcript.name, title)
-            cha_path = cha_path.replace(transcript.name, title)
-            transcript.name = title
-        reader.document.write_chat(cha_path)
-        # remove the old .txt transcript file
-        os.remove(transcript.content.path)
-        # set transcript file to .cha file
-        transcript.content = cha_name
+        if transcript.content.name.endswith('.txt'):
+            reader = SifReader(transcript.content.path)
+            cha_name = transcript.content.name.replace('.txt', '.cha')
+            cha_path = transcript.content.path.replace('.txt', '.cha')
+            # use title from metadata, if present
+            title = reader.document.title
+            if title:
+                cha_name = cha_name.replace(transcript.name, title)
+                cha_path = cha_path.replace(transcript.name, title)
+                transcript.name = title
+            reader.document.write_chat(cha_path)
+            # remove the old .txt transcript file
+            os.remove(transcript.content.path)
+            # set transcript file to .cha file
+            transcript.content = cha_name
+
+        elif transcript.content.name.endswith('.cha'):
+            pass
+
         transcript.status = 'converted'
         transcript.save()
         logger.info('Conversion successful')
